NAS-114558: Replaced combobox with combobox2
