+++ src/app/modules/ix-forms/classes/simple-combobox-provider.ts
@@ -1,6 +1,6 @@
 import { Observable, of } from 'rxjs';
 import { Option } from 'app/interfaces/option.interface';
-import { IxComboboxProvider } from 'app/modules/ix-forms/components/ix-combobox2/ix-combobox-provider';
+import { IxComboboxProvider } from 'app/modules/ix-forms/components/ix-combobox/ix-combobox-provider';
 
 export class SimpleComboboxProvider implements IxComboboxProvider {
   fetch(filterValue: string): Observable<Option[]> {
