+++ src/app/modules/ix-forms/components/ix-combobox/ix-combobox.component.html
@@ -20,29 +20,20 @@
       (blur)="onTouch()"
       [matAutocomplete]="auto"
     >
-    <span *ngIf="shouldShowResetInput()" class="reset-input">
-      <mat-icon (click)="resetInput()" role="img" fontSet="mdi-set" fontIcon="mdi-close-circle"></mat-icon>
+    <span *ngIf="shouldShowResetInput() || loading" class="reset-input">
+      <mat-progress-spinner *ngIf="loading" class="loader" mode="indeterminate" [diameter]="15"></mat-progress-spinner>
+      <mat-icon (click)="resetInput()" *ngIf="shouldShowResetInput()" role="img" fontSet="mdi-set" fontIcon="mdi-close-circle"></mat-icon>
     </span>
     <mat-autocomplete #auto="matAutocomplete" [displayWith]="getDisplayWith"
       (optionSelected)="optionSelected($event.option.value)" (opened)="onOpenDropdown()">
-
-      <ng-container *ngIf="filteredOptions | async as opts; else loadingOrError">
-        <mat-option
-          *ngFor="let option of filteredOptions | async"
-          [value]="option"
-          [class.ix-selected]="option.value === selectedOption?.value"
-          [class.ix-unselected]="option.value !== selectedOption?.value"
-        >
-          {{option.label}}
-        </mat-option>
-      </ng-container>
-
-      <ng-template #loadingOrError>
-        <ng-container *ngIf="errorObject">
-          <mat-option [disabled]="true">{{'Options cannot be loaded' | translate}}</mat-option>
-        </ng-container>
-      </ng-template>
-
+      <mat-option
+        *ngFor="let option of options"
+        [value]="option"
+        [class.ix-selected]="option.value === selectedOption?.value"
+        [class.ix-unselected]="option.value !== selectedOption?.value"
+      >
+        {{option.label}}
+      </mat-option>
     </mat-autocomplete>
   </div>
   <ix-errors [control]="controlDirective.control" [label]="label"></ix-errors>
