+++ src/app/modules/ix-forms/components/ix-combobox/ix-combobox.component.scss
@@ -40,7 +40,16 @@
   }
 }
 
+.loader {
+  display: inline-block;
+  margin-left: 4px;
+}
+
 .label-container {
+  align-items: center;
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-start;
   margin-bottom: 5px;
   margin-left: 1px;
 }
@@ -65,6 +74,7 @@
 .input-container mat-icon {
   font-size: 1rem;
   height: auto;
+  margin-left: 4px;
   width: auto;
 }
